Background: Use SVG
Use a SVG for the background image instead of a PNG. The new high-res,
scalable background is only 16KB compared to ~120KB.

Since the `bg.svg` file is the same as `low_bg.png` it has replaced that
background image as well. As a result, it is not necessary to include
b-lazy for lazy loading the images. Less JavaScript!